+++ rpc/backend/call_tx_test.go
@@ -6,6 +6,7 @@ import (
 	"math/big"
 
 	sdkmath "cosmossdk.io/math"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/rlp"
@@ -22,6 +23,7 @@ func (suite *BackendTestSuite) TestResend() {
 	gasPrice := new(hexutil.Big)
 	toAddr := tests.GenerateAddress()
 	chainID := (*hexutil.Big)(suite.backend.chainID)
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 	callArgs := evmtypes.TransactionArgs{
 		From:                 nil,
 		To:                   &toAddr,
@@ -67,6 +69,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFeeDisabled(queryClient)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce:   &txNonce,
@@ -89,6 +92,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce: &txNonce,
@@ -108,6 +112,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFeeDisabled(queryClient)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce:                &txNonce,
@@ -161,6 +166,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce:                &txNonce,
@@ -184,6 +190,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce:                &txNonce,
@@ -208,6 +215,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterParams(queryClient, &header, 1)
 				RegisterParamsWithoutHeader(queryClient, 1)
 				RegisterUnconfirmedTxsError(client, nil)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce:                &txNonce,
@@ -236,6 +244,7 @@ func (suite *BackendTestSuite) TestResend() {
 				RegisterParams(queryClient, &header, 1)
 				RegisterParamsWithoutHeader(queryClient, 1)
 				RegisterUnconfirmedTxsEmpty(client, nil)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				Nonce:                &txNonce,
@@ -437,6 +446,7 @@ func (suite *BackendTestSuite) TestDoCall() {
 
 func (suite *BackendTestSuite) TestGasPrice() {
 	defaultGasPrice := (*hexutil.Big)(big.NewInt(1))
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
 		name         string
@@ -456,6 +466,7 @@ func (suite *BackendTestSuite) TestGasPrice() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, sdkmath.NewInt(1))
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			defaultGasPrice,
 			true,
@@ -472,6 +483,7 @@ func (suite *BackendTestSuite) TestGasPrice() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, sdkmath.NewInt(1))
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			defaultGasPrice,
 			false,