+++ rpc/backend/sign_tx_test.go
@@ -35,8 +35,8 @@ func (suite *BackendTestSuite) TestSendTransaction() {
 		Gas:      &gas,
 		Nonce:    &nonce,
 	}
-
 	hash := common.Hash{}
+	validator := sdk.AccAddress(tests.GenerateAddress().Bytes())
 
 	testCases := []struct {
 		name         string
@@ -79,6 +79,7 @@ func (suite *BackendTestSuite) TestSendTransaction() {
 				RegisterBlock(client, 1, nil)
 				RegisterBlockResults(client, 1)
 				RegisterBaseFee(queryClient, baseFee)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			evmtypes.TransactionArgs{
 				From:     &from,
@@ -110,6 +111,7 @@ func (suite *BackendTestSuite) TestSendTransaction() {
 				txEncoder := suite.backend.clientCtx.TxConfig.TxEncoder()
 				txBytes, _ := txEncoder(tx)
 				RegisterBroadcastTxError(client, txBytes)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			callArgsDefault,
 			common.Hash{},
@@ -135,6 +137,7 @@ func (suite *BackendTestSuite) TestSendTransaction() {
 				txEncoder := suite.backend.clientCtx.TxConfig.TxEncoder()
 				txBytes, _ := txEncoder(tx)
 				RegisterBroadcastTx(client, txBytes)
+				RegisterValidatorAccount(queryClient, validator)
 			},
 			callArgsDefault,
 			hash,